split text stories (#3827)
